# public/openapi.yaml
openapi: 3.0.3
info:
  title: AgentsFlow API
  description: |
    API de automação de marketing via WhatsApp.

    ## Autenticação

    Todas as rotas protegidas exigem autenticação via:
    - **Session Cookie** (para navegador/UI)
    - **Bearer Token**: `Authorization: Bearer {AGENTSFLOW_API_KEY}`
    - **API Key Header**: `X-API-Key: {AGENTSFLOW_API_KEY}`

    ## Rotas Públicas

    As seguintes rotas NÃO exigem autenticação:
    - `/api/auth/*` - Login/logout
    - `/api/webhook` - Webhooks da Meta
    - `/api/health` - Health check
    - `/api/system` - Info do sistema
    - `/api/installer/*` - Setup inicial
  version: 1.0.0
  contact:
    name: AgentsFlow
  license:
    name: Proprietary
    url: https://agentsflow.app/license
servers:
  - url: http://localhost:3000
    description: Desenvolvimento
  - url: https://{domain}
    description: Produção
    variables:
      domain:
        default: seu-app.vercel.app
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Use AGENTSFLOW_API_KEY como token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Use AGENTSFLOW_API_KEY
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [Rascunho, Agendado, Enviando, Concluído, Pausado, Falhou, Cancelado]
        templateName:
          type: string
        recipients:
          type: integer
        sent:
          type: integer
        delivered:
          type: integer
        read:
          type: integer
        failed:
          type: integer
        createdAt:
          type: string
          format: date-time
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/campaigns:
    get:
      tags: [Campaigns]
      summary: Listar campanhas
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de campanhas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Campaigns]
      summary: Criar campanha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, templateName]
              properties:
                name:
                  type: string
                templateName:
                  type: string
                scheduledAt:
                  type: string
                  format: date-time
                contacts:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      name:
                        type: string
      responses:
        '201':
          description: Campanha criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
...